+++ src/main/java/com/onlyoffice/manager/request/DefaultRequestManager.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,6 +30,7 @@ import com.onlyoffice.model.settings.security.Security;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpStatus;
 import org.apache.http.StatusLine;
@@ -66,15 +67,18 @@ public class DefaultRequestManager implements RequestManager {
 
     /** {@link UrlManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final UrlManager urlManager;
+    @Setter(AccessLevel.PROTECTED)
+    private UrlManager urlManager;
 
     /** {@link JwtManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final JwtManager jwtManager;
+    @Setter(AccessLevel.PROTECTED)
+    private JwtManager jwtManager;
 
     /** {@link SettingsManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final SettingsManager settingsManager;
+    @Setter(AccessLevel.PROTECTED)
+    private SettingsManager settingsManager;
 
     /** {@link ObjectMapper}. */
     private final ObjectMapper objectMapper = new ObjectMapper();
@@ -143,7 +147,7 @@ public class DefaultRequestManager implements RequestManager {
                 StatusLine statusLine = response.getStatusLine();
                 if (statusLine == null) {
                     throw new ClientProtocolException(
-                            settingsManager.getSDKSetting("integration-sdk.product.name")
+                            settingsManager.getDocsIntegrationSdkProperties().getProduct().getName()
                                     + " URL: " + request.getURI() + " did not return a response.\n"
                                     + "Request: " + request.toString() + "\n"
                                     + "Response: " + response
@@ -153,7 +157,7 @@ public class DefaultRequestManager implements RequestManager {
                 HttpEntity resEntity = response.getEntity();
                 if (resEntity == null) {
                     throw new ClientProtocolException(
-                            settingsManager.getSDKSetting("integration-sdk.product.name")
+                            settingsManager.getDocsIntegrationSdkProperties().getProduct().getName()
                                     + " URL: " + request.getURI() + " did not return content.\n"
                                     + "Request: " + request.toString() + "\n"
                                     + "Response: " + response
@@ -163,7 +167,7 @@ public class DefaultRequestManager implements RequestManager {
                 int statusCode = statusLine.getStatusCode();
                 if (statusCode != HttpStatus.SC_OK) {
                     throw new ClientProtocolException(
-                            settingsManager.getSDKSetting("integration-sdk.product.name")
+                            settingsManager.getDocsIntegrationSdkProperties().getProduct().getName()
                                     + " URL: " + request.getURI() + " return unexpected response.\n"
                                     + "Request: " + request.toString() + "\n"
                                     + "Response: " + response.toString()
@@ -212,21 +216,21 @@ public class DefaultRequestManager implements RequestManager {
         Boolean ignoreSSLCertificate = settingsManager.isIgnoreSSLCertificate();
 
         Integer connectionTimeout = (int) TimeUnit.SECONDS.toMillis(
-                Long.parseLong(
-                        settingsManager.getSDKSetting("integration-sdk.http.client.request.connection.timeout")
-                )
+                settingsManager.getDocsIntegrationSdkProperties()
+                        .getHttpClient()
+                        .getConnectionTimeout()
         );
 
         Integer connectionRequestTimeout = (int) TimeUnit.SECONDS.toMillis(
-                Long.parseLong(
-                        settingsManager.getSDKSetting("integration-sdk.http.client.request.connectionRequest.timeout")
-                )
+                settingsManager.getDocsIntegrationSdkProperties()
+                        .getHttpClient()
+                        .getConnectionRequestTimeout()
         );
 
         Integer socketTimeout = (int) TimeUnit.SECONDS.toMillis(
-                Long.parseLong(
-                        settingsManager.getSDKSetting("integration-sdk.http.client.request.socket.timeout")
-                )
+                settingsManager.getDocsIntegrationSdkProperties()
+                        .getHttpClient()
+                        .getSocketTimeout()
         );
 
         if (httpClientSettings != null) {