+++ src/main/java/com/onlyoffice/manager/url/DefaultUrlManager.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,12 +20,19 @@ package com.onlyoffice.manager.url;
 
 import com.onlyoffice.manager.settings.SettingsManager;
 import com.onlyoffice.model.common.RequestedService;
+import com.onlyoffice.model.properties.docsintegrationsdk.DocumentServerProperties;
 import com.onlyoffice.model.settings.SettingsConstants;
+import com.onlyoffice.utils.ConfigurationUtils;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
-import java.text.MessageFormat;
+import java.beans.BeanInfo;
+import java.beans.IntrospectionException;
+import java.beans.Introspector;
+import java.beans.PropertyDescriptor;
+import java.lang.reflect.InvocationTargetException;
 
 
 @AllArgsConstructor
@@ -33,12 +40,13 @@ public class DefaultUrlManager implements UrlManager {
 
      /** {@link SettingsManager}. */
      @Getter(AccessLevel.PROTECTED)
-     private final SettingsManager settingsManager;
+     @Setter(AccessLevel.PROTECTED)
+     private SettingsManager settingsManager;
 
      @Override
      public String getDocumentServerUrl() {
          if (settingsManager.isDemoActive()) {
-              return sanitizeUrl(settingsManager.getSDKSetting("integration-sdk.demo.url"));
+              return sanitizeUrl(ConfigurationUtils.getDemoDocumentServerProperties().getUrl());
          } else {
               return sanitizeUrl(settingsManager.getSetting(SettingsConstants.URL));
          }
@@ -47,7 +55,7 @@ public class DefaultUrlManager implements UrlManager {
      @Override
      public String getInnerDocumentServerUrl() {
           if (settingsManager.isDemoActive()) {
-               return sanitizeUrl(settingsManager.getSDKSetting("integration-sdk.demo.url"));
+               return sanitizeUrl(ConfigurationUtils.getDemoDocumentServerProperties().getUrl());
           } else {
                String documentServerInnerUrl = settingsManager.getSetting(SettingsConstants.INNER_URL);
 
@@ -61,26 +69,66 @@ public class DefaultUrlManager implements UrlManager {
 
      @Override
      public String getDocumentServerApiUrl() {
-          return getDocumentServerUrl() + settingsManager.getSDKSetting("integration-sdk.api.url");
+          return getDocumentServerUrl() + settingsManager.getDocsIntegrationSdkProperties()
+                  .getDocumentServer()
+                  .getApiUrl();
      }
 
      @Override
      public String getDocumentServerPreloaderApiUrl() {
-          return getDocumentServerUrl() + settingsManager.getSDKSetting("integration-sdk.api.preloader.url");
+          return getDocumentServerUrl() + settingsManager.getDocsIntegrationSdkProperties()
+                  .getDocumentServer()
+                  .getApiPreloaderUrl();
      }
 
      @Override
      public String getServiceUrl(final RequestedService requestedService) {
+          String serviceUrl = null;
+
+          if (requestedService == null) {
+               return null;
+          }
+
           String serviceName = requestedService
                   .getClass()
                   .getInterfaces()[0]
                   .getSimpleName()
-                  .replaceAll("Service", "")
                   .toLowerCase();
 
-          String serviceUrl = settingsManager.getSDKSetting(
-                  MessageFormat.format("integration-sdk.service.{0}.url", serviceName)
-          );
+          DocumentServerProperties documentServerProperties = settingsManager.getDocsIntegrationSdkProperties()
+                  .getDocumentServer();
+
+          try {
+               Object serviceProperties = null;
+
+               BeanInfo beanInfo = Introspector.getBeanInfo(documentServerProperties.getClass());
+               for (PropertyDescriptor propertyDescriptor : beanInfo.getPropertyDescriptors()) {
+                    if (propertyDescriptor.getName().toLowerCase().equals(serviceName)) {
+                         serviceProperties = propertyDescriptor.getReadMethod().invoke(documentServerProperties);
+                    }
+               }
+
+               if (serviceProperties == null) {
+                    return null;
+               }
+
+               beanInfo = Introspector.getBeanInfo(serviceProperties.getClass());
+               for (PropertyDescriptor propertyDescriptor : beanInfo.getPropertyDescriptors()) {
+                    if (propertyDescriptor.getName().equals("url")) {
+                         serviceUrl = propertyDescriptor.getReadMethod().invoke(serviceProperties).toString();
+                    }
+               }
+          } catch (IntrospectionException e) {
+               throw new RuntimeException(e);
+          } catch (InvocationTargetException e) {
+               throw new RuntimeException(e);
+          } catch (IllegalAccessException e) {
+               throw new RuntimeException(e);
+          }
+
+          if (serviceUrl == null) {
+               return null;
+          }
 
           return getInnerDocumentServerUrl() + serviceUrl;
      }
@@ -92,6 +140,18 @@ public class DefaultUrlManager implements UrlManager {
           } else {
                return null;
           }
+     }
+
+     @Override
+     public String replaceToDocumentServerUrl(final String url) {
+          String documentServerUrl = getDocumentServerUrl();
+          String innerDocumentServerUrl = getInnerDocumentServerUrl();
+
+          if (!documentServerUrl.equals(innerDocumentServerUrl)) {
+               return url.replace(innerDocumentServerUrl, documentServerUrl);
+          }
+
+          return url;
      }
 
      @Override