+++ src/main/java/com/onlyoffice/service/documenteditor/config/DefaultConfigService.java
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -44,6 +44,7 @@ import com.onlyoffice.model.documenteditor.config.editorconfig.Template;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
 import java.util.List;
 import java.util.regex.Pattern;
@@ -54,19 +55,23 @@ public class DefaultConfigService implements ConfigService {
 
     /** {@link DocumentManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final DocumentManager documentManager;
+    @Setter(AccessLevel.PROTECTED)
+    private DocumentManager documentManager;
 
     /** {@link UrlManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final UrlManager urlManager;
+    @Setter(AccessLevel.PROTECTED)
+    private UrlManager urlManager;
 
     /** {@link JwtManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final JwtManager jwtManager;
+    @Setter(AccessLevel.PROTECTED)
+    private JwtManager jwtManager;
 
     /** {@link SettingsManager}. */
     @Getter(AccessLevel.PROTECTED)
-    private final SettingsManager settingsManager;
+    @Setter(AccessLevel.PROTECTED)
+    private SettingsManager settingsManager;
 
     @Override
     public Config createConfig(final String fileId, final Mode mode, final String userAgent) {
@@ -222,7 +227,10 @@ public class DefaultConfigService implements ConfigService {
     @Override
     public Type getType(final String userAgent) {
         Pattern pattern = Pattern.compile(
-                settingsManager.getSDKSetting("integration-sdk.mobile.user-agent"),
+                settingsManager.getDocsIntegrationSdkProperties()
+                        .getDocumentServer()
+                        .getEditingService()
+                        .getMobileUserAgent(),
                 Pattern.CASE_INSENSITIVE | Pattern.MULTILINE
         );