+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.0
+## Changed
+- list supported formats
+- improved work with Settings
+- UrlManager (added method replaceToDocumentServerUrl)
+- ar and sr empty file templates
+
 ## 0.0.1
 ## Added
 - common interfaces (DocumentManager, RequestManager, JwtManager, SettingsManager, UrlManager, ConfigService,